% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feo_read_nc.R
\name{feo_read_nc}
\alias{feo_read_nc}
\title{Reads fluxnetEO averaged netcdf data}
\usage{
feo_read_nc(file)
}
\arguments{
\item{file}{fluxnetEO file}
}
\value{
data frame
}
\description{
Given a fluxnetEO average file returns
a tidy data frame
}
