% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fluxnet_sites}
\alias{fluxnet_sites}
\title{Fluxnet Sites}
\format{
SpatialPolygonDataFrame
\describe{
  \item{site_code}{FLUXNET site code}
  \item{latitude}{latitude}
  \item{longitude}{longitude}
}
}
\usage{
fluxnet_sites
}
\description{
List of the location of fluxnet sites in the
FluxnetEO dataset
}
\keyword{datasets}
